Implement test coverage for edge functions
