+++ .github/workflows/tests.yml
@@ -39,3 +39,13 @@ jobs:
           flags: unittests
           name: codecov-umbrella
         continue-on-error: true
+
+  deno-test:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+      - uses: denoland/setup-deno@v2
+        with:
+          deno-version: v2.x
+      - name: Run edge function tests
+        run: cd supabase/functions && deno task test
