+++ supabase/config.toml
@@ -23,3 +23,9 @@ verify_jwt = false
 
 [functions.update-payment-method]
 verify_jwt = false
+
+[functions.generate-api-key]
+verify_jwt = false
+
+[functions.marketplace-dataset-urls]
+verify_jwt = false
