+++ supabase/functions/test/assert.ts
@@ -1,0 +1,17 @@
+export function assert(condition: unknown, message = "Assertion failed"): asserts condition {
+  if (!condition) throw new Error(message);
+}
+
+function stringify(value: unknown): string {
+  try {
+    return JSON.stringify(value);
+  } catch {
+    return String(value);
+  }
+}
+
+export function assertEquals(actual: unknown, expected: unknown, message?: string): void {
+  if (!Object.is(actual, expected)) {
+    throw new Error(message ?? `AssertionError: expected ${stringify(expected)}, got ${stringify(actual)}`);
+  }
+}
